Adding input validation

+++ Node/controllers/personalReward.js
@@ -4,6 +4,7 @@ const Reward = require("../models/Reward");
 const { randomString } = require("../utils/utils");
 const sequelize = require("../sequelize/sequelize");
 const { Sequelize } = require("sequelize");
+const { validationResult } = require("express-validator");
 
 /**
  * @swagger
@@ -58,56 +59,62 @@ const { Sequelize } = require("sequelize");
 
 
 module.exports.create = async (req, res) => {
-  if (req.session === undefined) {
-    return res.sendStatus(401);
-  }
-
-  const clientId = req.session.id;
-  const { reward_id } = req.body;
-
-  try {
-    const client = await Client.findByPk(clientId);
-    const reward = await Reward.findByPk(reward_id);
-
-    if (client === null)
-      throw new Error("Client not found");
-
-    if (reward === null)
-      throw new Error("Reward not found");
-
-    if (client.nb_throins < reward.throins_cost)
-      throw new Error("Too poor :'(");
-
-    await sequelize.transaction(
-      {
-        deferrable: Sequelize.Deferrable.SET_DEFERRED,
-      },
-      async (t) => {
-        await client.update({
-          nb_throins: client.nb_throins - reward.throins_cost
-        }, { transaction: t });
-
-        const discount_code = randomString();
-        let date = new Date();
-        date.setMonth(date.getMonth() + 2);
-        const exp_date = `${date.getFullYear()}-${date.getMonth()}-${date.getDate()}`;
-
-        const personalReward = await PersonalReward.create({
-          discount_code,
-          exp_date,
-          client_id: client.id,
-          reward_id: reward.id
+  const errors = validationResult(req);
+  if (!errors.isEmpty())
+    res.status(400).json({errors: errors.array() })
+  else {
+  
+    if (req.session === undefined) {
+      return res.sendStatus(401);
+    }
+  
+    const clientId = req.session.id;
+    const { reward_id } = req.body;
+  
+    try {
+      const client = await Client.findByPk(clientId);
+      const reward = await Reward.findByPk(reward_id);
+  
+      if (client === null)
+        throw new Error("Client not found");
+  
+      if (reward === null)
+        throw new Error("Reward not found");
+  
+      if (client.nb_throins < reward.throins_cost)
+        throw new Error("Too poor :'(");
+  
+      await sequelize.transaction(
+        {
+          deferrable: Sequelize.Deferrable.SET_DEFERRED,
         },
-          { transaction: t }
-        );
-        res.json(personalReward);
-      });
-  } catch (error) {
-    if (error.message === "Client not found" || error.message === "Reward not found" || error.message === "Too poor :'(")
-      res.status(404).json({ error: error.message });
-
-    console.error(error);
-    res.sendStatus(500);
+        async (t) => {
+          await client.update({
+            nb_throins: client.nb_throins - reward.throins_cost
+          }, { transaction: t });
+  
+          const discount_code = randomString();
+          let date = new Date();
+          date.setMonth(date.getMonth() + 2);
+          const exp_date = `${date.getFullYear()}-${date.getMonth()}-${date.getDate()}`;
+  
+          const personalReward = await PersonalReward.create({
+            discount_code,
+            exp_date,
+            client_id: client.id,
+            reward_id: reward.id
+          },
+            { transaction: t }
+          );
+          res.json(personalReward);
+        });
+    } catch (error) {
+      if (error.message === "Client not found" || error.message === "Reward not found" || error.message === "Too poor :'(")
+        res.status(404).json({ error: error.message });
+  
+      console.error(error);
+      res.sendStatus(500);
+    }
   }
 };
 /**
@@ -127,12 +134,19 @@ module.exports.create = async (req, res) => {
  *                              type: integer
  */
 module.exports.destroy = async (req, res) => {
-  const { id } = req.body;
-  try {
-    await PersonalReward.destroy({ where: { id } });
-    res.sendStatus(204);
-  } catch (error) {
-    console.error(error);
-    res.sendStatus(500);
+  const errors = validationResult(req);
+  if (!errors.isEmpty())
+    res.status(400).json({errors: errors.array() })
+  else {
+
+    const { id } = req.body;
+    try {
+      await PersonalReward.destroy({ where: { id } });
+      res.sendStatus(204);
+    } catch (error) {
+      console.error(error);
+      res.sendStatus(500);
+    }
+
   }
 };
